+++ library/ui/src/main/java/com/google/android/exoplayer2/ui/PlayerView.java
@@ -143,6 +143,12 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
  *         <li>Corresponding method: None
  *         <li>Default: {@code surface_view}
  *       </ul>
+ *   <li><b>{@code use_sensor_rotation}</b> - Whether to use the orientation sensor for rotation
+ *       during spherical playbacks (if available).
+ *       <ul>
+ *         <li>Corresponding method: {@link #setUseSensorRotation(boolean)}
+ *         <li>Default: {@code true}
+ *       </ul>
  *   <li><b>{@code shutter_background_color}</b> - The background color of the {@code exo_shutter}
  *       view.
  *       <ul>
@@ -306,6 +312,7 @@ public class PlayerView extends FrameLayout implements AdsLoader.AdViewProvider 
   @Nullable private Drawable defaultArtwork;
   private @ShowBuffering int showBuffering;
   private boolean keepContentOnPlayerReset;
+  private boolean useSensorRotation;
   @Nullable private ErrorMessageProvider<? super ExoPlaybackException> errorMessageProvider;
   @Nullable private CharSequence customErrorMessage;
   private int controllerShowTimeoutMs;
@@ -365,6 +372,7 @@ public class PlayerView extends FrameLayout implements AdsLoader.AdViewProvider 
     boolean controllerAutoShow = true;
     boolean controllerHideDuringAds = true;
     int showBuffering = SHOW_BUFFERING_NEVER;
+    useSensorRotation = true;
     if (attrs != null) {
       TypedArray a = context.getTheme().obtainStyledAttributes(attrs, R.styleable.PlayerView, 0, 0);
       try {
@@ -388,6 +396,8 @@ public class PlayerView extends FrameLayout implements AdsLoader.AdViewProvider 
                 R.styleable.PlayerView_keep_content_on_player_reset, keepContentOnPlayerReset);
         controllerHideDuringAds =
             a.getBoolean(R.styleable.PlayerView_hide_during_ads, controllerHideDuringAds);
+        useSensorRotation =
+            a.getBoolean(R.styleable.PlayerView_use_sensor_rotation, useSensorRotation);
       } finally {
         a.recycle();
       }
@@ -420,6 +430,7 @@ public class PlayerView extends FrameLayout implements AdsLoader.AdViewProvider 
         case SURFACE_TYPE_SPHERICAL_GL_SURFACE_VIEW:
           SphericalGLSurfaceView sphericalGLSurfaceView = new SphericalGLSurfaceView(context);
           sphericalGLSurfaceView.setSingleTapListener(componentListener);
+          sphericalGLSurfaceView.setUseSensorRotation(useSensorRotation);
           surfaceView = sphericalGLSurfaceView;
           break;
         case SURFACE_TYPE_VIDEO_DECODER_GL_SURFACE_VIEW:
@@ -741,6 +752,22 @@ public class PlayerView extends FrameLayout implements AdsLoader.AdViewProvider 
     if (this.keepContentOnPlayerReset != keepContentOnPlayerReset) {
       this.keepContentOnPlayerReset = keepContentOnPlayerReset;
       updateForCurrentTrackSelections(/* isNewPlayer= */ false);
+    }
+  }
+
+  /**
+   * Sets whether to use the orientation sensor for rotation during spherical playbacks (if
+   * available)
+   *
+   * @param useSensorRotation Whether to use the orientation sensor for rotation during spherical
+   *     playbacks.
+   */
+  public void setUseSensorRotation(boolean useSensorRotation) {
+    if (this.useSensorRotation != useSensorRotation) {
+      this.useSensorRotation = useSensorRotation;
+      if (surfaceView instanceof SphericalGLSurfaceView) {
+        ((SphericalGLSurfaceView) surfaceView).setUseSensorRotation(useSensorRotation);
+      }
     }
   }